+++ software/firmware/pru0-programmer/include/delay.h
@@ -2,14 +2,17 @@
 #define __DELAY_H_
 #include <stdint.h>
 
-#define F_CPU (200000000u)
+#define F_CPU            (200000000ul)
+#define TICK_INTERVAL_NS (5U)
 
-extern void        __delay_var_cycles(uint32_t cycles);
+static inline void delay_us(uint32_t time_us)
+{
+    while (time_us-- > 0u) __delay_cycles(1000ul / TICK_INTERVAL_NS);
+}
 
-static inline void delay_ns(uint32_t time_ns) { __delay_var_cycles(time_ns / 5u); }
-
-static inline void delay_us(uint32_t time_us) { __delay_var_cycles(time_us * 200u); }
-
-static inline void delay_ms(uint32_t time_ms) { __delay_var_cycles(time_ms * 200000ul); }
+static inline void delay_ms(uint32_t time_ms)
+{
+    while (time_ms-- > 0u) __delay_cycles(1000000ul / TICK_INTERVAL_NS);
+}
 
 #endif /* __DELAY_H_ */