+++ software/firmware/pru0-programmer/sbw_transport.c
@@ -41,7 +41,7 @@
 #include "delay.h"
 
 static gpio_state_t tclk_state = GPIO_STATE_LOW;
-static uint32_t     clk_delay_cycles;
+#define sbw_clk_delay_cycles (F_CPU / 500000ul / 2u)
 
 static struct
 {
@@ -50,30 +50,32 @@ static struct
     uint8_t sbw_dir;
 } pins;
 
+// TODO: tmsh to tdo_sbw were inline
+
 static void tmsh(void)
 {
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_HIGH);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_HIGH);
 }
 
 static void tmsl(void)
 {
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_HIGH);
 }
 
 static void tmsldh(void)
 {
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_HIGH);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_HIGH);
 }
@@ -81,35 +83,36 @@ static void tmsldh(void)
 static void tdih(void)
 {
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_HIGH);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_HIGH);
 }
 
 static void tdil(void)
 {
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_HIGH);
 }
 
 static gpio_state_t tdo_rd(void)
 {
-    gpio_state_t res;
     sys_gpio_cfg_dir(pins.sbw_tdio, GPIO_DIR_IN);
     sys_gpio_set(pins.sbw_dir, GPIO_STATE_LOW); // LOW => SWD_IO is INPUT
-    __delay_var_cycles(clk_delay_cycles);
+
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
-    res = sys_gpio_get(pins.sbw_tdio);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
+    const gpio_state_t res = sys_gpio_get(pins.sbw_tdio);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_HIGH);
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_HIGH);
     sys_gpio_set(pins.sbw_dir, GPIO_STATE_HIGH); // LOW => SWD_IO is INPUT
     sys_gpio_cfg_dir(pins.sbw_tdio, GPIO_DIR_OUT);
+    __delay_cycles(sbw_clk_delay_cycles);
 
     return res;
 }
@@ -118,18 +121,21 @@ static void tdo_sbw(void)
 {
     sys_gpio_cfg_dir(pins.sbw_tdio, GPIO_DIR_IN);
     sys_gpio_set(pins.sbw_dir, GPIO_STATE_LOW); // LOW => SWD_IO is INPUT
-    __delay_var_cycles(clk_delay_cycles);
+
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_LOW);
-    __delay_var_cycles(clk_delay_cycles);
+    __delay_cycles(sbw_clk_delay_cycles);
     sys_gpio_set(pins.sbw_tck, GPIO_STATE_HIGH);
     sys_gpio_set(pins.sbw_tdio, GPIO_STATE_HIGH);
     sys_gpio_set(pins.sbw_dir, GPIO_STATE_HIGH); // LOW => SWD_IO is INPUT
     sys_gpio_cfg_dir(pins.sbw_tdio, GPIO_DIR_OUT);
+
+    __delay_cycles(sbw_clk_delay_cycles);
 }
 
-void set_sbwtdio(gpio_state_t state) { sys_gpio_set(pins.sbw_tdio, state); }
+void set_sbwtdio(const gpio_state_t state) { sys_gpio_set(pins.sbw_tdio, state); }
 
-void set_sbwtck(gpio_state_t state) { sys_gpio_set(pins.sbw_tck, state); }
+void set_sbwtck(const gpio_state_t state) { sys_gpio_set(pins.sbw_tck, state); }
 
 void tmsl_tdil(void)
 {
@@ -256,7 +262,7 @@ int sbw_transport_init(const uint8_t pin_sbw_tck, const uint8_t pin_sbw_tdio,
 	 * In SLAU320AJ section 2.2.3.1., the 'delay' is specified as 5 clock cycles at 18MHz, but this seems
 	 * to not work reliably and contradicts the reference implementation.
 	 */
-    clk_delay_cycles = F_CPU / 500000ul / 2u;
+    // sbw_clk_delay_cycles = F_CPU / 500000ul / 2u;
 
     return 0;
 }